feat: refactored to separate files
