+++ src/error.rs
@@ -1,0 +1,1 @@
+pub type Result<T> = std::result::Result<T, anyhow::Error>;
